src/main/java/TicketManager.java
import java.util.Arrays;

public class TicketManager {
    private TicketRepository repo;

    public TicketManager(TicketRepository repo) {
        this.repo = repo;
    }

    public void add(Ticket ticket) {
        repo.save(ticket);
    }

    public Ticket[] searchBy(String from, String to) {
        Ticket[] result = new Ticket[0];
        for (Ticket ticket: repo.findAll()) {
            if (matches(ticket, from, to)) {
                Ticket[]tmp = new Ticket[result.length + 1];
                for (int i = 0; i < result.length; i++) {
                    tmp[i] = result[i];
                }
                tmp[tmp.length - 1] = ticket;
                result = tmp;
            }
        }
        Arrays.sort(result);
        return result;
    }

    public boolean matches(Ticket ticket, String search1, String search2) {
        if (ticket.getDepartureAirport().contains(search1) && ticket.getDestinationAirport().contains(search2)) {
            return true;
        } else {
            return false;
        }
    }
}
